chore(website): move and merge `eslint.config.js` to the root dir from the website dir
