+++ package.json
@@ -80,22 +80,23 @@
 		"commitizen": "4.3.0",
 		"cspell": "8.13.3",
 		"cz-conventional-changelog": "3.3.0",
+		"eslint-plugin-react": "7.35.0",
 		"heapdump": "0.3.15",
 		"husky": "9.1.4",
 		"lerna": "8.1.8",
 		"lint-staged": "15.2.9",
 		"npm-run-all2": "6.2.2",
-		"textlint": "14.0.5",
 		"textlint-filter-rule-comments": "1.2.2",
 		"textlint-rule-preset-ja-spacing": "2.4.3",
 		"textlint-rule-preset-ja-technical-writing": "10.0.1",
 		"textlint-rule-preset-japanese": "10.0.3",
 		"textlint-rule-preset-jtf-style": "2.3.14",
+		"textlint": "14.0.5",
 		"ts-node": "10.9.2",
 		"tsx": "4.17.0",
-		"typedoc": "0.26.5",
 		"typedoc-plugin-mdn-links": "3.2.8",
 		"typedoc-plugin-resolve-crossmodule-references": "0.3.3",
+		"typedoc": "0.26.5",
 		"typescript": "5.5.4"
 	},
 	"resolutions": {
