+++ website/package.json
@@ -38,7 +38,6 @@
     "@imgix/js-core": "3.8.0",
     "@tsconfig/strictest": "2.0.5",
     "@types/glob": "8.1.0",
-    "eslint-plugin-react": "7.35.0",
     "eslint": "9.9.0",
     "prettier": "3.3.3",
     "stylelint-config-recess-order": "5.0.1",
